+++ js/modules/bootstrap-select.js
@@ -14,26 +14,35 @@ $(function(){
     function initBootstrapSelect() {
         // dynamically define offset of select because of fixed site header
         let bs_select_offset_top = $('.site-header').outerHeight();
-        $('.selectpicker').selectpicker({
-            style: 'btn-default bs-placeholder',
-            size: 'auto',
-            dropupAuto: true, // per default, let bootstrap select choose where to dislpay the dropdown (above or below). Should you want to force to display the form below, add the `data-dropup-auto="false"` attribute to the desired <select> element.
-            windowPadding: [bs_select_offset_top,0,0,0],
-        });
-        setTimeout(function(){
-            $('.bootstrap-select .dropdown-toggle').each(function(){
-                this.innerHTML = this.innerHTML.replace(/&nbsp;/g,'');
-                $(this).find('.bs-caret').remove();
-                $(this).addClass('toggle-initialized');
-            });
-        },0);
+        let initialized_class = 'selectpicker--initialized';
+        $('.selectpicker, .selectpicker-alt').each(function(){
+            const $select = $(this);
+            if ($select.is(':visible') && $select.hasClass(initialized_class) === false) {
+                $select.selectpicker({
+                    style: 'btn-default bs-placeholder',
+                    size: 'auto',
+                    dropupAuto: true, // per default, let bootstrap select choose where to dislpay the dropdown (above or below). Should you want to force to display the form below, add the `data-dropup-auto="false"` attribute to the desired <select> element.
+                    windowPadding: [bs_select_offset_top,0,0,0],
+                });
+                // mark as initialized
+                $select.addClass(initialized_class);
+                // modify markup
+                setTimeout(function(){
+                    $select.parents('.bootstrap-select').find('.dropdown-toggle').each(function(){
+                        this.innerHTML = this.innerHTML.replace(/&nbsp;/g,'');
+                        // $(this).find('.bs-caret').remove();
+                        $(this).addClass('toggle-initialized');
+                    });
+                },0);
+            }
+        })
     }
 
     // on page load
     initBootstrapSelect();
 
     // custom event
-    $(window).on('initBootstrapSelect viewportWidthHasChanged ajaxForm:success softpage:opened form-modal:opened', function() {
+    $(window).on('initBootstrapSelect viewportWidthHasChanged ajaxForm:success softpage:opened form-modal:opened default-modal:opened', function() {
         initBootstrapSelect();
     });